chore: moved makeSwapErrorRight into utils
