+++ src/lib/swapper/constants.ts
@@ -9,7 +9,8 @@ import { thorchainSwapper } from 'lib/swapper/swappers/ThorchainSwapper/Thorchai
 import { zrxApi } from 'lib/swapper/swappers/ZrxSwapper/endpoints'
 import { zrxSwapper } from 'lib/swapper/swappers/ZrxSwapper/ZrxSwapper2'
 
-import { makeSwapErrorRight, SwapperName } from './api'
+import { SwapperName } from './api'
+import { makeSwapErrorRight } from './utils'
 
 export const QUOTE_TIMEOUT_MS = 10_000
