+++ src/lib/swapper/swappers/CowSwapper/utils/helpers/helpers.ts
@@ -10,7 +10,8 @@ import type { Asset } from 'lib/asset-service'
 import { bnOrZero, convertPrecision } from 'lib/bignumber/bignumber'
 import { fromBaseUnit } from 'lib/math'
 import type { SwapErrorRight } from 'lib/swapper/api'
-import { makeSwapErrorRight, SwapErrorType } from 'lib/swapper/api'
+import { SwapErrorType } from 'lib/swapper/api'
+import { makeSwapErrorRight } from 'lib/swapper/utils'
 
 import type { CowSwapQuoteResponse } from '../../types'
 import { CowNetwork } from '../../types'
