+++ src/lib/swapper/utils.ts
@@ -9,7 +9,25 @@ import type { FeatureFlags } from 'state/slices/preferencesSlice/preferencesSlic
 import { isCrossAccountTradeSupported } from '../../state/helpers'
 import { AsyncResultOf, isTruthy } from '../utils'
 import type { SwapErrorRight } from './api'
-import { makeSwapErrorRight, SwapErrorType, SwapperName } from './api'
+import { SwapErrorType, SwapperName } from './api'
+
+export const makeSwapErrorRight = ({
+  details,
+  cause,
+  code,
+  message,
+}: {
+  message: string
+  details?: unknown
+  cause?: unknown
+  code?: SwapErrorType
+}): SwapErrorRight => ({
+  name: 'SwapError',
+  message,
+  details,
+  cause,
+  code,
+})
 
 const getRequestFilter = (cachedUrls: string[]) => (request: Request) =>
   !cachedUrls.some(url => request.url.includes(url))
